requirements_web.txt
gradio>=4.0.0
opencv-python>=4.8.0
Pillow>=9.5.0
torch>=2.0.1
torchvision>=0.15.0
numpy>=1.24.0
diffusers==0.25.0
transformers==4.28.1
xformers==0.0.20
einops==0.7.0
open-clip-torch==2.20.0
peft==0.9.0
PyYAML==6.0
pyiqa==0.1.10
huggingface_hub==0.25.0
basicsr
loralib
fairscale
jax[cpu]==0.4.13
jaxlib[cpu]==0.4.13 